# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Service API
  version: 1.0.0
  description: REST API for Movie Service
  contact:
    name: Nikola Maravic
    email: n.maravic90@gmail.com
    url: https://www.linkedin.com/in/nmaravic/

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

tags:
  - name: MovieRead
    description: Public read operations (search, filter, pagination, rankings)
  - name: MovieCrud
    description: Create, update and delete operations (Requires ADMIN role)
  - name: MovieImage
    description: Movie image management (Requires ADMIN role)
  - name: MovieReview
    description: Movie reviews and ratings (Requires USER role)

paths:

  /movies:
    get:
      tags: [MovieRead]
      summary: List of movies
      description: Returns a paginated list of movies with optional filters and sorting.
      operationId: getMovies
      parameters:
        - name: query
          in: query
          description: Search by title or description
          required: false
          schema:
            type: string
            example: Inception
        - name: genre
          in: query
          description: Filter by genre
          required: false
          schema:
            $ref: '#/components/schemas/Genre'
        - name: year
          in: query
          description: Filter by release year
          required: false
          schema:
            type: integer
            example: 2010
        - name: minRating
          in: query
          description: Minimum average rating
          required: false
          schema:
            type: number
            format: double
            minimum: 1.0
            maximum: 10.0
            example: 7.5
        - name: sort
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [title, releaseYear, avgRating]
            default: title
        - name: order
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successfully returned movie list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePageResponse'

    post:
      tags: [MovieCrud]
      summary: Create movie
      description: Creates a new movie. Requires ADMIN role.
      operationId: createMovie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieRequest'
            example:
              title: Inception
              overview: A thief who steals corporate secrets through dream-sharing technology.
              releaseYear: 2010
              director: Christopher Nolan
              genres: [SCI_FI, THRILLER]
      responses:
        '201':
          description: Movie successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Movie with the same title and year already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /movies/top-rated:
    get:
      tags: [MovieRead]
      summary: Top rated movies
      description: Returns a list of movies sorted by average rating, highest first.
      operationId: getTopRatedMovies
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of top rated movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieResponse'

  /movies/{movieId}:
    get:
      tags: [MovieRead]
      summary: Movie details
      operationId: getMovieById
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '200':
          description: Movie found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [MovieCrud]
      summary: Replace entire movie
      description: |
        Full replace - all fields are required.
        Fields not sent become null.
        Requires ADMIN role.
      operationId: replaceMovie
      parameters:
        - $ref: '#/components/parameters/MovieId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieRequest'
      responses:
        '200':
          description: Movie successfully replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [MovieCrud]
      summary: Partial movie update
      description: |
        Updates only the fields provided - others remain unchanged.
        Requires ADMIN role.
      operationId: updateMovie
      parameters:
        - $ref: '#/components/parameters/MovieId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoviePatchRequest'
            example:
              overview: Updated overview text.
              genres: [SCI_FI, ACTION]
      responses:
        '200':
          description: Movie successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [MovieCrud]
      summary: Delete movie
      description: |
        Deletes the movie and all related data (CASCADE DELETE).
        Requires ADMIN role.
      operationId: deleteMovie
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '204':
          description: Movie successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /movies/{movieId}/images:
    post:
      tags: [MovieImage]
      summary: Upload movie image
      description: |
        Upload an image for a movie.
        Exactly one COVER image is allowed.
        Maximum 3 SLIDE images allowed.
        Requires ADMIN role.
      operationId: uploadMovieImage
      parameters:
        - $ref: '#/components/parameters/MovieId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - type
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP - max 5MB)
                type:
                  $ref: '#/components/schemas/ImageType'
      responses:
        '201':
          description: Image successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieImageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cover already exists or maximum slide limit (3) reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [MovieImage]
      summary: Get movie images
      description: Returns all images for a movie.
      operationId: getMovieImages
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '200':
          description: List of movie images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieImageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [MovieImage]
      summary: Delete all movie images
      description: Deletes all images for a movie. Requires ADMIN role.
      operationId: deleteAllMovieImages
      parameters:
        - $ref: '#/components/parameters/MovieId'
      responses:
        '204':
          description: All images successfully deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /movies/{movieId}/reviews:
    post:
      tags: [MovieReview]
      summary: Add review
      description: |
        Adds a review and rating for a movie.
        One review per user per movie.
        Requires USER role.
      operationId: createReview
      parameters:
        - $ref: '#/components/parameters/MovieId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
            example:
              rating: 9
              comment: One of the best movies ever made.
      responses:
        '201':
          description: Review successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User has already reviewed this movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [MovieReview]
      summary: Get movie reviews
      description: Returns a paginated list of reviews for a movie.
      operationId: getMovieReviews
      parameters:
        - $ref: '#/components/parameters/MovieId'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /movies/{movieId}/reviews/{reviewId}:
    delete:
      tags: [MovieReview]
      summary: Delete review
      description: |
        User can delete their own review.
        ADMIN can delete any review.
      operationId: deleteReview
      parameters:
        - $ref: '#/components/parameters/MovieId'
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Review successfully deleted
        '403':
          description: Not allowed to delete this review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:

  parameters:
    MovieId:
      name: movieId
      in: path
      required: true
      description: Movie ID
      schema:
        type: integer
        format: int64
        example: 1

  schemas:

    Genre:
      type: string
      enum:
        - ACTION
        - COMEDY
        - DRAMA
        - HORROR
        - SCI_FI
        - THRILLER
        - CRIME
        - ROMANCE
        - DOCUMENTARY
        - ANIMATION
        - FANTASY
        - MYSTERY
        - ADVENTURE
        - WAR
        - BIOGRAPHY
        - SPORT
        - WESTERN

    ImageType:
      type: string
      enum:
        - COVER
        - SLIDE

    MovieRequest:
      type: object
      required:
        - title
        - releaseYear
        - director
        - genres
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Inception
        overview:
          type: string
          maxLength: 2000
          example: A thief who steals corporate secrets through dream-sharing technology.
        releaseYear:
          type: integer
          minimum: 1888
          maximum: 2100
          example: 2010
        director:
          type: string
          maxLength: 255
          example: Christopher Nolan
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
          minItems: 1

    MoviePatchRequest:
      type: object
      description: All fields optional — only provided fields are updated.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        overview:
          type: string
          maxLength: 2000
        releaseYear:
          type: integer
          minimum: 1888
          maximum: 2100
        director:
          type: string
          maxLength: 255
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'

    MovieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Inception
        overview:
          type: string
          example: A thief who steals corporate secrets through dream-sharing technology.
        releaseYear:
          type: integer
          example: 2010
        director:
          type: string
          example: Christopher Nolan
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        avgRating:
          type: number
          format: double
          example: 8.8
        ratingCount:
          type: integer
          example: 42
        images:
          type: array
          items:
            $ref: '#/components/schemas/MovieImageResponse'
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date

    MoviePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MovieResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 150
        totalPages:
          type: integer
          example: 8

    MovieImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        url:
          type: string
          example: https://storage.example.com/movies/1/cover.jpg
        type:
          $ref: '#/components/schemas/ImageType'
        createdAt:
          type: string
          format: date

    ReviewRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 9
        comment:
          type: string
          maxLength: 1000
          example: One of the best movies ever made.

    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        movieId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: nmaravic
        rating:
          type: integer
          example: 9
        comment:
          type: string
          example: One of the best movies ever made.
        createdAt:
          type: string
          format: date

    ReviewPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReviewResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 150
        totalPages:
          type: integer
          example: 8

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Movie with id 99 not found
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/movies/99

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'